+++ services/catalog-api/jest.config.js
@@ -7,4 +7,5 @@ module.exports = {
   collectCoverageFrom: ['**/*.(t|j)s'],
   coverageDirectory: '../coverage',
   testEnvironment: 'node',
+  transformIgnorePatterns: ['/node_modules/(?!jose/)'],
 };