+++ services/catalog-api/src/modules/auth/auth.module.ts
@@ -5,8 +5,11 @@ import { CommonAccessModule } from '../../common/common-access.module';
 import { AuthController } from './auth.controller';
 import { AuthIdentityService } from './auth-identity.service';
 import { AuthService } from './auth.service';
+import { AppleAuthLoginService } from './social-auth/apple-auth-login.service';
+import { AppleIdentityTokenVerifierService } from './social-auth/apple-identity-token-verifier.service';
 import { GoogleAuthLoginService } from './social-auth/google-auth-login.service';
 import { GoogleIdTokenVerifierService } from './social-auth/google-id-token-verifier.service';
+import { SocialAuthLoginService } from './social-auth/social-auth-login.service';
 
 @Module({
   imports: [
@@ -26,9 +29,18 @@ import { GoogleIdTokenVerifierService } from './social-auth/google-id-token-veri
   providers: [
     AuthService,
     AuthIdentityService,
+    SocialAuthLoginService,
     GoogleIdTokenVerifierService,
     GoogleAuthLoginService,
+    AppleIdentityTokenVerifierService,
+    AppleAuthLoginService,
   ],
-  exports: [JwtModule, AuthService, AuthIdentityService, GoogleAuthLoginService],
+  exports: [
+    JwtModule,
+    AuthService,
+    AuthIdentityService,
+    GoogleAuthLoginService,
+    AppleAuthLoginService,
+  ],
 })
 export class AuthModule {}